🌈 style(app): UI 大改

+++ src/actions/pwd.ts
@@ -1,33 +1,86 @@
 'use server'
 
 import { db } from '@/db'
-import { passwords, PasswordUpdateDTO } from '@/db/schema/passwords'
-import { users } from '@/db/schema/users'
-import { eq } from 'drizzle-orm'
+import { PasswordInsertDTO, passwords, PasswordUpdateDTO } from '@/db/schema/passwords'
+import { aesDecrypt, aesEncrypt } from '@/utils/aes-encrypt.server'
+import { and, eq, getTableColumns } from 'drizzle-orm'
+import { isNil, omit } from 'es-toolkit'
+import { getUserInfo } from './auth'
 
-export async function createPassword(formData: FormData) {
-  const userList = await db.select().from(users)
+const ENCRYPTION_KEY = process.env.ENCRYPTION_KEY || ''
+const generateIV = () => Math.random().toString(36).slice(2, 10)
 
-  const name = formData.get('name') as string
-  const email = formData.get('email') as string
+export async function getUserPasswords() {
+  const currentUser = await getUserInfo()
 
-  if (userList.length === 0) {
-    throw new Error('No user found')
-  }
+  const cols = omit(getTableColumns(passwords), ['password', 'iv'])
+  const passwordsList = await db.select(cols).from(passwords).where(eq(passwords.uid, currentUser.id))
+  return passwordsList
+}
 
+export async function createPassword(data: PasswordInsertDTO) {
+  const currentUser = await getUserInfo()
+
+  const { password, ...rest } = data
+  const iv = generateIV()
+  const encrypted = aesEncrypt(password || '', ENCRYPTION_KEY, iv)
   await db.insert(passwords).values({
-    uid: userList[0].id,
-    title: name,
-    username: email,
-    password: `${name}_${email}`,
-    iv: '123123'
+    ...rest,
+    uid: currentUser.id,
+    password: encrypted,
+    iv
   })
 }
 
 export async function updatePassword(params: PasswordUpdateDTO) {
-  await db.update(passwords).set(params).where(eq(passwords.id, params.id))
+  const currentUser = await getUserInfo()
+
+  const { id, password, ...rest } = params
+  const [data] = await db
+    .select()
+    .from(passwords)
+    .where(and(eq(passwords.id, id), eq(passwords.uid, currentUser.id)))
+
+  if (!data) {
+    throw new Error('Password not found')
+  }
+
+  const encrypted = aesEncrypt(password || '', ENCRYPTION_KEY, data.iv)
+  const updated = { ...rest, password: encrypted }
+
+  await db
+    .update(passwords)
+    .set(updated)
+    .where(and(eq(passwords.id, id), eq(passwords.uid, currentUser.id)))
 }
 
 export async function removePassword(id: string) {
+  const currentUser = await getUserInfo()
+
+  const [data] = await db
+    .select()
+    .from(passwords)
+    .where(and(eq(passwords.id, id), eq(passwords.uid, currentUser.id)))
+
+  if (isNil(data)) {
+    throw new Error('Password not found')
+  }
+
   await db.delete(passwords).where(eq(passwords.id, id))
+}
+
+export async function getPasswordDetail(id: string) {
+  const currentUser = await getUserInfo()
+
+  const [data] = await db
+    .select()
+    .from(passwords)
+    .where(and(eq(passwords.id, id), eq(passwords.uid, currentUser.id)))
+
+  if (isNil(data)) {
+    throw new Error('Password not found')
+  }
+
+  const decrypted = aesDecrypt<string>(data.password || '', ENCRYPTION_KEY, data.iv || '')
+  return omit({ ...data, password: decrypted }, ['iv'])
 }
